GH-345: [refactor] Initial refractor of events lists lists in the profile page (created and joined) now have a feed that passes the API function to match the homepage feed
